move counter access test to an own test.
